NOV-52 Fine grained scope check via annotation for DropWizard resources
Provide the @FineGrainedScopeAllowed annotation to protect resources in
DW against fine grained scopes which can depend on the request context,
i.e. a fine grained scope can be used to match request parameters and
deny the request if the parameter value is not the correct one.

The dynamic matching is not completely free form and expects the scopes
be formatted using the fine grained scope approach to try to mitigate
errors that can lead to security issues.