Set default cassandra health check query as idempotent
The default health check query is not set as idempotent, which can might
end up marking an instance as unhealthy in situations where they should
still be able to function. This change will set it to idempotent to fix
those situations.

Also, allow setting the health check query idempotence in the
configuration when a custom query is used. If a custom query is used and
idempotence is not specified, fail CassandraConfiguration validation
checks.